+++ .prettierrc
@@ -1,0 +1,9 @@
+{
+	"tabWidth": 2,
+	"singleQuote": false,
+	"jsxSingleQuote": false,
+	"arrowParens": "avoid",
+	"semi": false,
+	"trailingComma": "none",
+	"useTabs": true
+}
